+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Changed
 - Remove digests from ICO files.
+- Remove ModDate and ID from PDF files.
 
 ## [1.1.0] - 2021-01-18
 ### Changed